requirements.txt
selenium==4.8.0
PyYAML~=6.0